Created ClientServer.java (#137)
* Created ClientServer.java

ClientServer.java gets request from client and activates the senddatatoclient method in proto. It initializes all the parts of the API, gets them to do the computation, then it returns a message back to the client telling them whether it was successful and what the output type is. (I havent actually put in any error messages yet)

* get rid of extra ;